various reorganization for alchemical_gromacs.py
* reorganized files to reduce duplication.  Now only one alchemical_analysis files.
* parsers moved to their own folder.
* unixlike commands are moved to their own module to reduce duplication.
* added __init__.py files make the imports work better
* README.md's are rewritten to put the complete option documentation into
  the main README, with gromacs and sire specific commands in the README.md's
